Move SharePosition and OptionPosition to modules.
Give them their own modules so that there's much less clutter in trying to
define a record type. This also renames position.hs to Position.hs because it
seems Haskell has the same idea as Java where a file name must also match the
module name.

The gentle introduction to Haskell claims modules and files are decoupled, but
I haven't gotten that to work so far.